fix: move lib/auth/session.ts into server folder
since it uses some env vars, incorrect to put it in lib
